+++ feature-libs/product-configurator/rulebased/cpq/common/index.ts
@@ -5,5 +5,5 @@
  */
 
 export * from './converters/index';
-export * from './cpq-configurator-common.modules';
+export * from './cpq-configurator-common.module';
 export * from './cpq.models';
