+++ feature-libs/product-configurator/rulebased/cpq/rulebased-cpq-configurator.module.ts
@@ -6,7 +6,7 @@
 
 import { NgModule } from '@angular/core';
 import { provideDefaultConfig } from '@spartacus/core';
-import { CpqConfiguratorCommonModule } from './common/cpq-configurator-common.modules';
+import { CpqConfiguratorCommonModule } from './common/cpq-configurator-common.module';
 import { defaultConfiguratorCpqConfig } from './config/default-configurator-cpq.config';
 import { CpqConfiguratorOccModule } from './occ/cpq-configurator-occ.module';
 import { CpqConfiguratorRestModule } from './rest/cpq-configurator-rest.module';
